buildsys: detect matching Lua and tolua++ pairs
So far, we have tried to find the most recent Lua version on the system
and then a matching tolua++. Since we have only recently provided a
patch for Lua 5.2 support in tolua++, most systems have not been
upgraded, yet. Therefore, it makes more sense to look for a Lua version
for which we can find corresponding Lua and tolua++ version, e.g.
resulting in the use of Lua 5.1 on lab machines unless our custom
package is installed. This is better than not having Lua code available
at all.