openprs: wait for server to come up on plugin init
When the plugin is initialized, wait for some (configurable) time for
the OPRS-server to be reachable. This is relevant when starting on a
heavily loaded system were the process might actually require some time
to be fully operational. Otherwise a plugin using the OpenPRSAspect
loaded directly afterwards might fail since the server is not yet
available. It als ensures that when disabling handling of the server
that we check if a server is actually reachable.