openprs: wait for kernel to come up for some time
Similarly to a server coming up taking some time (previous commit)
kernels are also external processes requiring time. We waited half a
second until now, but that may not be enough. We now have a configurable
timeout that we wait for the kernel to have connected to the server
proxy. Then we can be assured that transmitting commands will work.