libnavgraph: move navgraph bits from utils to own library
The topological map graph bits have been moved to their own library.
From ticket #345:

The code for the navgraph is currently spread over the utils library
and the navgraph plugin. Especially with the new constraints feature
things got worse because the constraint library resides in the navgraph
plugin.

The idea is to factor out the graph and constraints libraries into a
consistent library. Then, the navgraph plugin will use those libraries
(utils will no longer have navgraph related code). We can even consider
moving the NavGraphAspect to the library as a sub-library.

This also accounts also for the still increasing importance of the
navgraph on which many critical features (skills, agent, navigation)
build. It will stand out more and be easier to recognize.

This commit is the first step for the journey.