libnavgraph: move NavGraphAspect and constraints to library
The aspect continues to be a separate library as it it not directly
related to the navgraph itself. But it makes sense to have it at the
same place to make navgraph-using plugins independent from the default
navgraph plugin.

Move the constraints into the main navgraph library as we expect many
more inter-dependencies going forward as it becomes an integral part.