libnavgraph: make constraint repo intrinsic navgraph property
The constraint repository directly related to the navgraph and is only
useful in its context. Additionally, whenever planning to make search a
feature of the NavGraph -- something required to allow searching outside
of the navgraph plugins, e.g. for cost estimation -- the NavGraph must
always have access to the constraint repository.

Therefore we now make the constraint repo a member and intrinsic
property of the navgraph. The aspect has been adapter such that there is
no extra constraint_repo field anymore, but rather it must be queried
from the navgraph.

Along the way, the ConstraintRepo has been renamed to
NavGraphConstraintRepo to make its function more explicit.