core: set the waiting flag of a thread to false when thread is cancelled
If a thread is cancelled it will not run again, thus it cannot be
waiting for wakeup. This partially solves an issue with the thread list,
which would try to recover the thread even if it was already cancelled,
leading to garbled output and possibly other issues.

This is related to #315.