core: check if a thread is cancelled during thread recovery
If a thread is flagged as bad, the thread list will try to recover the
thread. Recovery does not check whether a thread is already cancelled,
it also recovers threads which are cancelled. This causes reads on freed
memory.

Check if the thread is cancelled and if so, remove it from the barrier
without doing anything else with it.

This fixes #315.