libnavgraph: move search into NavGraph and generalize
The NavGraph now provides methods to perform a graph search. There it
supports overriding the cost and estimate functions with custom version,
or to use default/registered ones. This now allows for a plugin to
override the cost and estimate functions for the A* search, e.g allowing
to replace the default distance-based method using, for example, actual
learned or simulated travel time information.

The search result is now a NavGraphPath instance, which combines several
utility features currently used in the navgraph plugin.

This change makes it possible to search for paths from anywhere now and
not only from within the navgraph plugin. This opens up new
possibilities to make decisions based on path queries.