libnavgraph: add Voronoi-based navgraph generator
Introduced the concept of a navgraph generator and implemented on based
on Voronoi diagrams. This generator computes a navgraph from a number of
given obstacles. The Voronoi diagram ensures that the graph always keeps
the maximum possible distance to all surrounding obstacles.

The generator is rather simple and does not provide many sanity checks
or constraints at the moment (a bounding box check is included).