navgraph: make navgraph file parameter optional
With generators being available it now can make sense to load the
navgraph plugin without a pre-existing graph. A generator can create one
later in another plugin. The navgraph plugin is still required as the
provider of the navgraph and to use the navgraph to drive.

We may consider splitting these two concerns into separate plugins at a
later point in time.