navgraph: visualization coalesces wakeups and locks graph
Make sure to lock the graph to avoid race conditions leading to
segfaults (when modifying the graph). Additionally coalesce wakeups. We
might be triggered by a number of changes like node insertions, each
triggering a wakeup. But the publishing cannot run because the lock
cannot (yet) be acquired. Then later publishing the graph numerous times
is harmful (lots of traffic).