core: release the WaitCondition's internal mutex if thread is cancelled
The WaitCondition locks its mutex before it wait()s if it uses an
internal mutex. The lock is automatically released in the wait call and
reaquired after wait() returns. However, if the thread is cancelled while
it is waiting, the mutex is locked before the thread terminates. Thus,
any other threads using the same WaitCondition will end up in a
deadlock.

Note that we only release the mutex if an internal mutex is used. If the
WaitCondition uses an external mutex, it is the caller's task to release
the mutex on thread cancellation.

This fixes #346.