core: add ThreadLoopListener
With the ThreadLoopListener, it is possible to do tasks right before the
thread's loop() starts and right after the thread's loop() returned.
In order to do so, you need to inherit from ThreadLoopListener,
implement the pre_loop and post_loop methods and register your loop
listener with the respective thread.