cmake: add cmake modules to use Fawkes for external code
These cmake modules allow to find and use a Fawkes tree in an external
project. These files were originally developed as part of a YAGI
application in the Logistics League domain based on Fawkes.

The modules allow for the inclusion of some core libraries directly
supported by the main file and are extensible in that there can be
component-specific cmake modules, e.g. to determine pre-requisite
availability. Examples are navgraph and tf for which this commit
contains appropriate modules.