config: fix barrier race condition in network config
When mirroring the config on a network config, a barrier is used to
synchronize receiving the message with initial data and the calling
thread. However, the calling thread may already destroy the barrier
while the other one is still waiting to be woken up. The barrier has
been properly reached by both, but since the thread deleting the barrier
is called just a tiny bit earlier than the other one, this later one
will then segfault.