laser_lines: post-process lines to filter large gaps
Even though we set the max sample distance for drawing samples, once
the line coefficients have been determined it takes all points which
could "support" the model, that is all points whose distance is closer
than a given threshold. But this means that when detecting points say on
some furniture item and then on the other side of the room at a wall it
will still use these points for the line, causing large jitter.

We now post-process a line by looking for the largest consistent cluster
within the line support points therefore eliminating large gaps and
distant points. That drastically improves stability and makes the length
calculation accurate.

This commit is a backport from changes made to the ROS laser_lines port.