laser-lines: calc_lines optional parameter to get remaining points
This adds a new optional paramter to calc_lines() that allows to get the
remaining point cloud after all points which are accounted to a line
have been removed. This can be useful to use line segmentation as a
pre-processing step or if the remainder can still contain useful
information.