dynamixel: generic Dynamixel servo driver plugin
This plugin exposes individual interfaces for configured servo chains.
Servos are automatically detected on the bus.
This plugin is based on the pantilt plugin's Dynamixel driver originally
developed for the RX28-based PTU. It is meant as a more generic hardware
driver for the Dynamixel servos. This allows to place servos with
different semantics, e.g. gripper and a camera tilt servo, on the same
chain without having a plugin that has to mix these functionalities.