core: do not protect pre and post loop hooks with loop mutex
The loop mutex should only protect the actual loop but not the pre and
post loop hooks. Now, we can interrupt (and cancel) a thread which
is currently in a pre or post loop hook.  As an example, we can cancel a
thread which is currently waiting for a syncpoint in the pre loop.
If the pre loop is protected with the mutex, cancelling the thread would
not be possible.