eclipse-clp: move externals to plugin directory
The eclipse externals aren't a generic library and are not used by any
other component, thus move the files to the plugin directory.

The externals are now directly linked into the plugin, no extra library
is created. This avoids issues with linking, as part of the externals
depend on the EclipseAgentThread, while the EclipseAgentThread depends
on parts of the externals.