dynamixel: add alarm_shutdown to DynamixelServoInterface
Dynamixel servos turn torque off, initiate an alarm-shutdown
and set a one-byte-length bitmask indicating the reason what failed.
The possible errors are as follows (taken from Dynamixel AX12 Documentation):

Bit 7: 0
Bit 6: If set to 1, torque off when an Instruction Error occurs
Bit 5: If set to 1, torque off when an Overload Error occurs
Bit 4: If set to 1, torque off when a Checksum Error occurs
Bit 3: If set to 1, torque off when a Range Error occurs
Bit 2: If set to 1, torque off when an Overheating Error occurs
Bit 1: If set to 1, torque off when an Angle Limit Error occurs
Bit 0: If set to 1, torque off when an Input Voltage Error occurs

The alarm_shutdown can now be read from DynamixelServoInterface from the
field alarm_shutdown.