dynamixel: add raw error code information from servo
Dynamixel servos can respond with error flags set in the error-byte given in
a status packet (this has to be configured for the particular servo with
the status_return_level - the dynamixel plugin automatically sets all available
servos to send a status packet only on READ commands).

The error flags indicate the following errors:
Instruction Error, Overload Error, Checksum Error, Range Error,
Overheating Error, Angle Limit Error and Input Voltage Error.

The dynamixel plugin sets these error flags in a field called "error" in
the interface by performing a bitwise-or with the currently sensed
errors. This is needed because the servo doesn't seem keep an error
state on alarm shutdown.
The error-field in the interface can be reset by sending a
ResetRawError-Message to the interface - the new value will then be 0.