libnavgraph: allow specifying the insertion mode in graph file
The insertion mode for a node can now be controlled using a property in
a graph file. For example a node may be connected to the closest node or
edge. The connection is created after all nodes have already been added.
Multiple connection insertions are performed in the order specified in
the graph file. Hence a later connection to an edge might use a
previously automatically created one.