clips: set bool in interface(msg) Type wrong fix
Somehow when setting a boolean field to the CLIPS::Value resulting from
the CLIPS term TRUE the type is TYPE_STRING instead of TYPE_SYMBOL. This
looks like a bug in clipsmm but I did not investigate this further.
Because it makes no difference, I also allowed the type string here.