dynamixel: add prevent_alarm_shutdown functionality
Dynamixel servos tend to initiate their alarm_shutdown when the current
load is about 40% or more of the configured torque_limit and the
alarm_shutdown-flags are set to at least 0x32 (shutdown on overload error)
resulting in a temporarily non-working servo.
As just disabling this flag can be dangerous for the motor, windings and
gears of the servo the new prevent_alarm_shutdown functionality releases
the load of the servo by moving its angle towards the direction of the
load until the measured load is below a configurable value
(prevent_alarm_shutdown_threshold, 0 - 1) of the configured torque_limit.