dynamixel: add min- and maximum allowed voltage in config
The plugin used to fail initialization in case of a voltages outside of the range
from 12V to 16V which wasn't the best option in case batteries with a rather low
voltage were used.
The values are now to be set as min_voltage and max_voltage in the
configuration.