dynamixel: add disable or enable prevent_alarm_shutdown
The DynamixelServoInterface now supports a field and a message called
SetPreventAlarmShutdown which is used to disable or enable the
prevent_alarm_shutdown feature of the particular servo managed by the
dynamixel plugin.
This is needed in case the user wants to temporarily ignore the overload
condition of the particular servo.