dynamixel: enable_connection_stability in servo_chain
The newly implemented connection_stability feature makes sure that there
is data available to read on the particular servo chain when the
recv-method is executed.
This makes sure that even if there is a connection drop to the
servo-chain, the plugin doesn't block indefinitely and can continue the
communication when the chain is available again.