buildsys: target subdirs only needs to depend on current target
The target subdirs should not always depend on LIBS_all (BINS_all etc.)
but only on the libs (bins etc.) of the current target. If the current
target is 'gui', then subdirs should only depend on LIBS_gui (BINS_gui
etc.). Similarly for the target presubdirs.
This fixes a bug where everything is built for the 'gui' target even if
the gui target itself is empty.