navgraph: timeout only for distance,send final ori-msg if req (fixes #349)
The timeout is used as initially intended: Allow local navigator
to get closer to the target, before we stop it.
The "orientation" property of the node is not affected by it
anymorei This means: If the timeout fires and the node has an
"orientation" property, we send a final "goto-node" command with
x=y=0, i.e. "only rotate". Otherwise we can stop explicitly.

Naturally, if the local navigator becomes final in the meantime,
navgraph becomes final as well.

In the "rotate only" phase we have no rotation-timeout yet/anymore.
In this case the navgraph is final only when colli is final. We
do have a check if the orientation-tolerance of the navgraph is
fulfilled, but no action for that case yet.