navgraph: add timeout after orientation has been reached
If a "rotate only" command is sent after the last node has
been reached, and the orientation falls within the tolerance
at some point after, we start a new timer that stops the
local navigator in case it takes too long to become final.
It is basically the same timeout-behaviour we have for the
target-tolerance.
See #349 for further details.