syncpoint: correctly manage pending components of predecessor syncpoints
In general, the predecessor syncpoint should be emitted whenever a
syncpoint is emitted. However, if the current component was not pending
(i.e. it has already emitted the syncpoint), then the component should
not be removed from the pending components of the predecessor syncpoint,
and all components which wait_all() for the predecessor should not be
unblocked, but the predecessor should still be emitted.

This allows a component to register for multiple syncpoints and emit a
single syncpoint multiple times. As an example, say the component 'c1'
is registered as emitter for '/test/sp1' and '/test/sp2', and thus,
because of the syncpoint hierarchy, it is also registered as emitter for
'/test' twice, once for each successor syncpoint. Now c1 emits
'/test/sp1' and since it was pending, it is removed from the pending
components for both '/test/sp1' and '/test'. Then, if c1 *again* emits
'/test/sp1', it is not pending anymore. Thus, it should not be removed
from the pending components of '/test' (since now it is only pending
because of the second syncpoint). If it were removed, then components
which wait_all() for '/test' would be unblocked even though '/test/sp2'
has never been emitted.