libwebview: support setting TLS cipher suite
So far webview always used the default cipher suite which is rather
broad and also allows for no longer recommended ciphers and protocols
like SSLv3. The new default cipher suite is rather restricted to only
secure ciphers with 128 bits or more (stream cipher) based on TLS 1.2.
SSLv3 and prior to 1.2 are disabled. This can be overridden if
necessary.