config: explicitly set errno to zero before calling strtol
The recent problem of misinterpreting a config value was caused by errno
carrying an invalid value -- invalid in the sense that it was set by
something before calling strtol. strtol does not set errno explicitly to
zero if there is no error. That is, for older glibc versions at least.
Since we had no issues on more recent versions I guess that might have
changed. But documentation on the web suggests this is optional and not
a requirement.