lua: support additional init and loop funcs for skill jumpstate
The base jump state supports setting a number of init or loop functions.
This is required for timeouts to work. Since they were not supported in
a SkillJumpState (due to a bad design in the way the hierarchy is
implemented, which I shall fix WhenThereIsTime(TM)) the timeouts always
immediately fired.