libnavgraph: allow multi graphs during search_path()
When the reachability calulations were deferred to the next
search_path() it would enforce connectedness. However, at that point we
should make the least assumptions about the graph and hence allow a
multigraph. If the caller does not intend so he can still call
calc_reachability() directly with this not allowed.

Therefore add missing parameter for lesser constraints.