dynamixel: add discovery of explicit set servos
The plugin now lets the user define a specific set of servos to look for
during the discovery phase.
This can be set by e.g. servos: [1,2,3] in dynamixel.yaml for the given
chain. When the user doesn't provide a set of dynamixel servos in the
configuration (servos: []), the broadcasting method is used which sends a PING instruction to
all available servos.
An exception is thrown if the plugin doesn't find a desired servo in the
chain.