dynamixel: add config handling for explicit servos
The dynamixel.yaml - config lets the user define a set of servos to be
discovered (e.g. servos: [1,2,3]) rather than exploring all available
servos.
This functionality is useful in the case of connection problems with
dynamixel servos that can't be discovered by sending a PING-instruction
to the broadcast.
If the user prefers sending pings via broadcast the servos-field in the
configuration should be left blank (servos: []).