skiller: fix passing subskill argument in its 'args' field (fixes #364)
Subskill arguments that are passed in the form
   self.skills[1].args = {...}
are now added properly to the subskill's arguments table.
They are also resetted properly.

In case the user sets any other field of the subskill in the form of
   self.skills[1].some_field = ...
, this is still added as an argument to not break existing code
(see #364 and #363), but a warning will be shown.

An additional condition has been added, so that setting
   self.skills[1].status
will be ignored.