laser-filter: by default reset data to quiet NaN
The value used for resetting the output buffer was 0.0. This is fine
when assuming laser data. However, if the data is further processed,
e.g. as a point cloud, this assumption may not generally hold (for point
clouds) and therefore cause problems.

Now, set the output buffer to a quiet NaN by default.