jaco: enable planning for angular targets
Previously, angular targets were not processed by OpenRAVE and
passed directly to libkindrv (when processed from the queue).
This could lead to problems, for example when we want to move
to a known "safe" position for which we have the joint angles
(like the angles for the "RETRACT" position). Now we can get
to those target positions with collision-avoidance.