joystick: add safety lockout feature
We have the issue that some joysticks set an initial (wrong) value when
started until move a bit. This can cause the robot to unexpectedly move
at the beginning.

The joystick will only write data to the blackboard now after receiving
an "unlock" combo, which is holding any button and moving the two
primary axes all the way to min and max, i.e. make a circle with the
primary cross. After completing the combo and releasing the key again
the safety lockout is lifted.