core: release the loop listeners' mutex when joining the thread
In order to allow proper finalization, the loop listeners' mutex must be
unlocked in all cases while the thread is joined. Otherwise, if the
mutex is still locked, any aspect using a loop listener cannot finalize,
as the aspect tries to remove itself from the loop listeners during
finalization, leading to a deadlock. Such a deadlock occurs especially
often if the loop listener's pre_loop and post_loop calls do not return
immediately (e.g. if pre_loop waits for a syncpoint).

Unlocking the mutex after joining the thread is safe because the loop
listeners will not be called anymore and the list of loop listeners will
not be accessed otherwise.