syncpoint: add the possibility of time-limited waits
Sometimes a component should only wait for a certain time for a given
syncpoint. After the time limit has been exceeded, the component will
print a warning and continue to execute.

In particular, this is used in the main thread to accomplish timed waits
for the hooks of the main loop.