libconfig: unlock mutex after successful emit
Mutex was only unlocked in case of some exception, not
when successfully finishing the emit.
Beforehand the config could only be written once and
additional set_value calls would not have any effect.

As I understood it, the write_pending_ variable should
have ensured that everything is written to file on exit.
Since information was lost, there might be an additional
bug here.