buildsys: disable deprecation warning for CLIPS
When compiling modules that use CLIPS there are warnings regarding a
deprecated std::auto_ptr. That is because we by default enable C++11
support for CLIPS stuff, which warns about this now. The warning is
triggered by glibmm, which is in turned used by clipsmm.

To get around this (until glibmm is fixed) we disable deprecated
declaration warnings for CLIPS-related code.