buildsys: adapt libxml++ detection to glibmm 2.46
Starting with version 2.46 glibmm has the mandatory requirement of
C++11. Since this is neither the default for GCC nor is the appropriate
flag included in the pkg-config build flags, we must take care to add it
ourselves. Move libxml++ bits into an own separate file out of
interface.mk to make it easer to re-use.