buildsys: fix generated build flags for ROS Jade
ROS Jade creates library link commands of the form
"-l:/full/path/to/lib.so" in the generated pkg-config files. This used
to work with older ld version, but does not longer with recent versions.
Therefore, adopt a workaround discussed for rosbuild by removing "-l:",
as ld allows to simply name shared libraries on the command line.

Confer:
https://github.com/ros/catkin/issues/694
https://github.com/ros/ros/pull/87