tf: upgrade fawkestf to ROS tf2
The tf library has been upgraded with a port of ROS tf2. Rather than
creating a tf2 library, we opt to simply replace the current library.
This is akin to ROS, where the tf API simply forwards to the tf2 API.

A general requirement for tf2 comes from the fact that we want to be
able to fully integrate with newer ROS versions. This in particular
requires a special handling of static transforms, which are assumed to
be perpetually valid once published. This makes computations more
efficient on typical systems with several static transforms.

We make a principal change on the Fawkes side in how transforms are
stored in the blackboard. We now expect there to exist one
TransformInterface per frame ID. With the tree structure, that means
that each frame is published with respect to another reference frame.
The child frame is what defines an interface instance. Static transforms
are indicated with a boolean flag in the data, rather than with some
name pattern. This makes it much easier (or rather possible) to inspect
tf data in the blackboard (e.g., via webview).

A fundamental difference to ROS is that we rely only on tf data types
(and its underlying bullet data types). ROS tf2 unfortunately uses a
message type as data exchange format, which cannot be ported with much
wider ROS dependencies. With our approach, we keep the API almost
stable, and we still get the benefits of better maintained code and
higher performance.

Note that frame IDs may no longer start with a slash. That was done in
ROS tf2 because of dropping the tf prefix (which we never used).

We have applied similar changes to adapt the ROS code to Fawkes as done
before for tf1. C++11 is now a mandatory requirement for tf. We have
backported some tools from tf1 (e.g., dot graph generation) to keep API
stable and for simplicity.