libconfig: add node path to exception message in YamlConfiguration
When a config value is rejected by the config parser because it doesn't
match the hard-coded constraints for that value type, it may be
difficult to find the responsible config file. With this, we output the
full YAML path of the config value so the user can at least have a rough
idea of where to look.