netcomm: fix deadlock and memory corruption in server thread
The clients map was locked extensively. This could cause a deadlock, for
example in the case that an interface was opened locally at the same
time it was closed remotely. This happened frequently, e.g., in the RCLL
simulation when starting with gdb enabled. Fawkes would startup more
slowly and AMCL initialization would overlap with ffset_pose.

Additionally, handlers were not locked properly and could be corrupted.
This, however, is more a theoretical excercise at this point.