ros-tf: avoid boost::bind to fix tf self-filtering
This problem manifests in many messages that a transform cannot be
published because a writer has already been opened. This happens because
the incoming messages (over the ROS topic) are not properly filtered, as
the publisher name is always "unknown_publisher". Therefore, whenever
sending a local transform to ROS, we receive it ourselves and cannot
properly filter it.

Using boost::bind to set a parameter whether the received transform is
static breaks the connection header of message events. My suspicion is
that the event will not be passed by reference but as a copyable (as the
_1 placeholder must support the CopyConstructible concept) and it seems
as if the connection header is not properly transferred.

Just avoiding boost::bind altogether fixes this problem.