skiller: add new SkillJumpState as parent class for skill-executing classes
As discussed in #368 and in person during team-meetings, this
introduces "SkillJumpState" as a parent implementor class for all
other types of states that can execute a skill (e.g. the
AgentSkillExecJumpstate for luaagent; or SubskillJumpState for
skiller, which was "SkillJumpState" until now).

This makes it easier for developers of skills or luaagents, for they
can use similar syntax in their files, i.e. use "SkillJumpState"
for a state that executes skills (although the internal procedures
are totally different, depending on the implemented class)