libinterface: improve Lua wrapper typing
The methods to access messages returned the generic instance of a
message. This can be improved by supporting returning the specific type
of the message to avoid having to always cast. This is akin to the
recent blackboard wrapper.

Along the way, add some missing methods to the message wrapper.