liblua: add support for finalization calls
Some Lua context may need specific finalization. Additionally, when
restarting a context (e.g., because of file changes) it might be
necessary to free resources in the old context, so that the new one can
be initialized. But if initializing the new context fails, they must be
re-aquired. For this also introduce finalization preparation and
cancellation calls.