skiller: support dynamic interface allocation
This removes the need to configure used interfaces in the configuration,
but rather opens interfaces during initialization. This is a long
standing issue where we now catch up to the ROS port which did this for
quite some time. The capability to use the config has been completely
removed.

This will need further changes to properly embed this, e.g., the
LuaInterfaceImporter is no longer necessary and read and write calls
must be issued at appropriate times.