skiller: implement run-time fully in Lua
This is the first step to remove C++ code that performs initialization,
communication, and run-time coordination for the behavior engine. This
patch implements all of this functionality in Lua only and requires
only little support from the surrounding C++ code. An update to the
skiller plugin is pending to make use of this new Lua code base.

The code has been aligned with the ROS port of the behavior engine. It
would benefit from another refactoring cycle but it overall makes it
much easier to follow the code path in the skiller already.