skiller: use Lua-based run-time
This change removes most of the handling code of the skiller which is
now embedded as pure Lua. This makes it much easier to update and
maintain the skiller and the code clearer to understand.

The remaining code is mainly setting up the basic environment and
pre-loading some packages. This can probably be reduced further in a
later step.