buildsys: extend tolua++ generation with exception handling code
Use the tolua++ code generator hooks to add support for Fawkes and std
exception handling. Catch the appropriate exceptions and convert them
into a Lua error. This avoids messing up Lua's error handling longjmp
chain and invalid stacks. It also allows to detect certain problems more
gracefully, e.g. if a message is sent to an interface without a writer
this will be handled on the Lua level as a regular Lua error.

This is also a requirement for a more robust (almost) pure Lua skiller.