skiller: support re-loading with writing interfaces
In the (almost) pure Lua version, interfaces are maintained within Lua
and not in the surrounding C++ code (through LuaInterfaceImporter). This
causes a problem when initializing a new context during a (file event
induced) reload: the new context would try to open the very same
interfaces, which fails for writing instances, as there may be at most
one such instance.

Solve this by re-sorting to the C++ code to transfer the writing
interface instances to the new context and to make sure that these are
not closed if initialization fails (e.g., in a skill initialized later).