libconfig: fix GCC optimization problem
There was a problem that GCC 5.3 seems to apply some optimization which
breaks the loop copying the values. Fix by making the loop counter
volatile and make compiler being strict about loads and stores. Not sure
what the actual cause of the problem is but proposed fix works for me.