libplugin: better error handling on config errors in plugin ctors
If a plugin loaded through a meta plugin threw an exception due to a
config value missing required in the ctor, it was mis-interpreted as a
"no meta plugin by this name" causing it to try to load a library of the
meta plugin name, which again failed. The true cause of the failure was
thus hidden and hard to debug.

Fix by explicitly checking for the config value first.