robotino: re-factor blackboard access for sensor data
The sensor thread now holds the set of sensor interfaces instead of the
com thread. For this clear separation we accept the intermediate data
structure. This eliminates code duplication among different com thread
implementations.

Along the way, fix a bug that caused the OpenRobotino com thread to run
much more often (virtually in a busy loop), due to the cycle time being
specified in milliseconds but the tool requiring microseconds...