robotino: fill in method bodies, full communication working
This is the first version to fully allow for sensor reading and setting
motor commands. Joystick control looks just fine.

A modified communication scheme might be useful for more efficient error
recovery more akin to controld3, i.e., just sending a "ping" message
and parsing it instead of closing and re-opening the serial connection.
This would require to remove the coupling of send_and_recv, which seems
feasible in our case.