ros-pcl: add bidirectional exchange of pointclouds
the ros-pcl plugin just copied pcs from fawkes->ros
now it also sends pcs from ros-fawkes

For this a polling (every 10 sec) is done on the ros topics to check for
available topics. If the first pointcloud from ROS is received, this topic will be
available in fawkes, one cloud needs to be received since otherwise the
Point-Type used in this topic is unknown and a known type is a requirement for the
pcl_manager. The supported Point-Types are:
- XYZ
- XYZRGB (untested)
- XYZI (untested)
Furthermore only topics that are used in fawkes are subscribed in ros,
for this also a polling is done, which results in an offset up to 10 sec
untill the first cloud is received in fawkes.