ros-pcl: fixed segmentation fault (workaround?)
if a pointcloud (fawkes->ros) is used in ros and the sender in fawkes is
closed, the adapter in ros-pcl was closed, but the adapter was allready
deleted during the closing in the sender.

The best solution would be a fix in the adapter class, however it seems
to look like a check is implemented, I couldn't find the reason why this
does not work.