skiller: when multiple exec_skill msgs are enqueued use last one
When multiple exec_skill messages are enqueued, the skiller has to
decide which one to use. Now we use the last one instead of the first
one because this is a similar behavior compared to calling multiple
skills with a small time step in between so the skiller starts executing
the first and switches to the second when receiving the message a later
loop iteration.

This also solves some issues in the CLIPS agent calling two skills to
fast without a timeout in between.