clips-protobuf: destroy messages on cleanup
Messages were not properly destroyed when no longer used resulting in a
memory leak. The protobuf-msg facts were retrated, but the related
pointer was not destroyed. Thanks to Frederik Zwilling for catching
this!

There used to be automatic cleanup code in the very first iteration of
this code. It registered to the periodic signal and used a map of
asserted facts for each message. If a fact was retracted (and thus no
longer referenced by anyone else) the message could be deleted. The code
had not been updated when moving this code into its own library already
back in the llsf-refbox (in fact, the handle_periodic_signals() is still
there in the refbox).