ffifacegen: generate packed interface structs
Interface structures used to be created 4-byte packed. However, changing
this to a (1-byte) packed layout has two advantages:
- we generate smaller structs reducing the network load (somewhat)
- the packing/alignment is more predictable

Especially the latter drove me to this change. Compilers are virtually
free to choose their in-struct alignment. The problem in combination
with jfawkes was that single-byte ints (int8/uint8) no longer were
padded. The deserialization of TestInterface data therefore failed due
to a different expectation in terms of length.

The one minor disadvantage is that computation speed could suffer --
would we ever compute directly on interface values. However, if strict
aligntment had been a requirement, it would have been risky/broken
before already, 4-byte alignment on a 64-bit machine doesn't help much
here.

Did I mention that it is time for a blackboard 2.0 with better data
structures and marshalling?