robot-memory: registering, checking, and executing computables
Computables provide information computation on demand when querying
(e.g. for interfacing perception, and geting complex information that
would be inpractical to store permanently in the robot memory such as
object distances). You can register a computable in the robot memory
by giving a function that returns a list of computed documents. The
function receives the query that should be executed at the moment and
which usually contains more information about what should be computed.
When registering a computable, you also need to give an identifyer
query that defines when to call the computable by checking if the
current query is matched by the identifyer query.