worldmodel: remove (soccer) world model plugin
The world model plugin was specific to robot soccer in the RoboCup
Middle Size and Standard Platform (Nao) Leagues. The fusers might still
be useful but may then be recovered from version control. They have not
been used for quite a while.

The name was highly misleading as it really was a soccer-specific
version of a world model, and even in that sense biased towards the
mentioned leagues and the specific behavior specifications. It was not a
generic world model that would have justified that plugin name.