netcomm: proper scope IDs for IPv6 addresses
For IPv6 it is more common to have prefix overlap, especially for
link-local addresses (fe80: prefix) of multiple network devices.
This causes connect() to fail with an "invalid argument" error. Fix this
by determining the interface through which the connection would be
established (translate interface index to name), create a prefix string
of the form IP%device, and run this through getaddrinfo to get a
properly setup sockaddr_in6 struct. Pass that along to event receivers.