gui_utils: service chooser connect directly to discovered endpoint
Now that the service model can record the proper sockaddr structure use
that as endpoint when connecting to the remote. This has a few benefits.
Mostly, we profit from the proper interface scope ID determination in
the avahi thread, additionally we gain a little speed avoiding going
through another lookup step.