ros-navgraph-breakout: allow global path planning and execution in ROS
This plugin provides the same execution API as the navgraph, but then
calls a specific ROS action for planning and execution to a specified
point of interest. In combination with the navgraph plugin with disabled
execution this may still be based on the existing navgraph, e.g., as
generated by navgraph-generator or statically defined.