navgraph: expose ROS services for path search and pairwise costs
This extends the ROS interface (publishing the navgraph) by a service to
search for a path between two nodes (or from the current position or to
a freely chosen goal position). Additionally, a service allows to
calculate pairwise costs among a number of nodes. This can be a typical
inquiry for task planning systems for cost estimation and is handled
more efficiently this way than with many (hundreds or more) calls to the
path search service.