robotino: only publish odom transform if enabled in configuration
This adds a new config value
/hardware/robotino/odometry/publish_transform. Only publish a transform
odom->base_link if the config value is set to true. This allows us to
use other software such as ROS' robot_localization for computing the
odometry while still accessing the Robotino sensors.

If the config value is not found, default to true to preserver the old
behavior.