amcl: catch tf exception in case the published transform is invalid
In certain cases, the transform computed by AMCL is invalid, which leads
to an exception when publishing the transform. In that case, catch the
exception and print an error. This results in not up-to-date AMCL
transforms, but avoids a termination due to an uncaught exception.

This kind of error occurs occassionaly when using ROS'
robot_localization for the odometry. Without catching the exception,
AMCL often causes a termination due to the exception after a few
seconds.