protobuf_comm: gracefully handle name resolution errors
The peer is stateless in that there is no connection that has been
established. However, if a host name (rather than an address) is given
as the receiving endpoint, it must be resolved. In some situations, this
may fail initially and for a limited time, e.g., when running in a
container cluster such as Kubernetes and the peers might be other pods.

We still want to keep the name resolution asynchronous (therefore, the
ctor cannot just fail if name resolution fails) and we would like to get
this with minimal changes to other code. Therefore, this is implemented
as a feature internal to the broadcast peer. If resolution fails, it
will keep trying in two second intervals. Through the send error signal,
callers can now about this and still destroy the peer.